more fixes
-revert schema.sql changes
-remove prettier config
-remove esling config
-remove automatic tax in stripe
-fix subscription query return
