+++ app/account/page.tsx
@@ -31,12 +31,12 @@ export default async function Account() {
 
   return (
     <section className="mb-32 bg-black">
-      <div className="mx-auto max-w-6xl px-4 py-8 sm:px-6 sm:pt-24 lg:px-8">
+      <div className="max-w-6xl px-4 py-8 mx-auto sm:px-6 sm:pt-24 lg:px-8">
         <div className="sm:align-center sm:flex sm:flex-col">
           <h1 className="text-4xl font-extrabold text-white sm:text-center sm:text-6xl">
             Account
           </h1>
-          <p className="m-auto mt-5 max-w-2xl text-xl text-zinc-200 sm:text-center sm:text-2xl">
+          <p className="max-w-2xl m-auto mt-5 text-xl text-zinc-200 sm:text-center sm:text-2xl">
             We partnered with Stripe for a simplified billing.
           </p>
         </div>
@@ -51,10 +51,9 @@ export default async function Account() {
           }
           footer={<ManageSubscriptionButton session={session} />}
         >
-          <div className="mb-4 mt-8 text-xl font-semibold">
+          <div className="mt-8 mb-4 text-xl font-semibold">
             {subscription ? (
-              `${subscriptionPrice}/$
-            {subscription?.prices?.interval}`
+              `${subscriptionPrice}/${subscription?.prices?.interval}`
             ) : (
               <Link href="/">Choose your plan</Link>
             )}
@@ -72,12 +71,12 @@ export default async function Account() {
             </div>
           }
         >
-          <div className="mb-4 mt-8 text-xl font-semibold">
+          <div className="mt-8 mb-4 text-xl font-semibold">
             <form id="nameForm">
               <input
                 type="text"
                 id="name"
-                className="w-1/2 rounded-md bg-zinc-800 p-3"
+                className="w-1/2 p-3 rounded-md bg-zinc-800"
                 defaultValue={userDetails?.full_name}
                 placeholder="Your name"
                 maxLength={64}
@@ -90,7 +89,7 @@ export default async function Account() {
           description="Please enter the email address you want to use to login."
           footer={<p>We will email you to verify the change.</p>}
         >
-          <p className="mb-4 mt-8 text-xl font-semibold">
+          <p className="mt-8 mb-4 text-xl font-semibold">
             {user ? user.email : undefined}
           </p>
         </Card>
@@ -108,13 +107,13 @@ interface Props {
 
 function Card({ title, description, footer, children }: Props) {
   return (
-    <div className="p m-auto my-8 w-full max-w-3xl rounded-md border border-zinc-700">
+    <div className="w-full max-w-3xl m-auto my-8 border rounded-md p border-zinc-700">
       <div className="px-5 py-4">
         <h3 className="mb-1 text-2xl font-medium">{title}</h3>
         <p className="text-zinc-300">{description}</p>
         {children}
       </div>
-      <div className="rounded-b-md border-t border-zinc-700 bg-zinc-900 p-4 text-zinc-500">
+      <div className="p-4 border-t rounded-b-md border-zinc-700 bg-zinc-900 text-zinc-500">
         {footer}
       </div>
     </div>
